Add syslinux-themes-squibby
Besides graphics stuff, syslinux-themes-squibby is how we implement boot support for
Greek keyboard, so I decided to put the package on this same "core" project for
squibby. Seeds and other packages will be in their own repositories.

For now syslinux-themes-squibby actually contains kubuntu graphics / branding
but this is just for testing and will change very soon.